move functional detail out of detail subfolder
