+++ Samples/CMakeLists.txt
@@ -3,6 +3,7 @@ set(HEADERS
     SampleHandlerInterface.h
     SampleHandlerBase.h
     SampleHandlerBase.txx
+    SampleHandlerFunctional.h
     SampleStructs.h
     HistogramUtils.h
     SampleInfo.h
@@ -35,7 +36,6 @@ install(TARGETS Samples
         LIBRARY DESTINATION lib/
         PUBLIC_HEADER DESTINATION include/Samples)
 
-install(DIRECTORY detail DESTINATION include/Samples)
 
 add_library(MaCh3::Samples ALIAS Samples)
 target_link_libraries(MaCh3 INTERFACE Samples)
