+++ Samples/SampleHandlerFunctional.h
